[FLINK-13650][tests] Consolidate CommonTestUtils and ClassLoaderUtils
