+++ flink-core/src/test/java/org/apache/flink/api/java/typeutils/runtime/kryo/KryoSerializerClassLoadingTest.java
@@ -22,7 +22,7 @@ import org.apache.flink.api.common.ExecutionConfig;
 import org.apache.flink.api.common.typeutils.SerializerTestBase;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.api.java.tuple.Tuple1;
-import org.apache.flink.core.testutils.CommonTestUtils;
+import org.apache.flink.testutils.ClassLoaderUtils;
 
 import org.junit.After;
 import org.junit.Before;
@@ -39,8 +39,8 @@ import static org.junit.Assert.fail;
 public class KryoSerializerClassLoadingTest extends SerializerTestBase<Object> {
 
 	/** Class loader and object that is not in the test class path. */
-	private static final CommonTestUtils.ObjectAndClassLoader<Serializable> OUTSIDE_CLASS_LOADING =
-			CommonTestUtils.createSerializableObjectFromNewClassLoader();
+	private static final ClassLoaderUtils.ObjectAndClassLoader<Serializable> OUTSIDE_CLASS_LOADING =
+			ClassLoaderUtils.createSerializableObjectFromNewClassLoader();
 
 	// ------------------------------------------------------------------------
